changed to BOOTSTRAP
I did it.
I started using Bootstrap for this website.

So far, it;s producing MOSTLY good results. There are a lot of "templates" given such as the jumbotron, row and column manager, image resizer, and more.

The only difficult thing is getting the PDF to size properly. Sometimes when going in inspect mode, the pdf is set to a low zoom like 20%. But in default, desktop mode it;s fine.

I copied and made back-ups for the pre-bootstrap html and css for the index