Fixed top navbar
That's it. A small victory but still a victory nonetheless. Previously, the non-bootrsap-compliant hmtl and css for the navbar, would always be behind the bootstrap elements.
Now the nav-bar has been "bootstrapified" and works.
It's also a lot cleaner AND WORKS almost flawlessly on mobile.

I need to add the dropdown menu button thing