Added some of the projects page content
I add some of the content for the Projects page.
it's not fully complete, as the one project has assets that can't be accessed. The other is just in a less-usual format and so i've asked the client how I should present it.

I added another css class to separate the main heading/title from the paragraph or smaller heading/titles.
The main title will be used at the header of the container, while the smaller titles will be used inside the jumbotrons.