Changed https to hpp + updated readme
I updated the read me, to have the backup link incase the desired link has 404 errors.

I also tried a small change which changes https to htpp for the imported/linked stylesheets and scripts.
The website has http in its link, so i'm guessing the http's need to be the same