Cleaned up some code, assets, and added the .nojekyll
The website works perfectly on desktop but not on mobile (or atleast the first page. I tested on my phone, and the other pages had css working except for index.html)

I decided to clean up some of the html and css, removing the unnecessary comments that had old paths. I also removed images that aren't being used.

lastly, i added the .nojekyll page. This is something i'm not too sure about but I hope it works for fixing mobile view. I'm not using / nor plan on using jekyll