fixed favicons on later pages + update 1st page links
i fixed the favicons on the later pages. they had the wrong file path xd

I also updated a few of the links that link stylesheet, JS, bootstrap to have https instead of http. This is bevause the mobile version shows issues on the first page, but the other pages work flawlessly