Replaced images + added image description
I checked the website this morning and saw that the two images on the hero-page, which had a source linked to LinkedIn, were not being displayed. Clicking on the images directed you to the linkedin, but the actual images weren't showing.
So i downloaded the images and just added them to the project folder. This shouldn't impact the optimization and current speed of the website (I hope it doesn't)

I also added a few descriptions (alt="" in HTML), to the images on the front-page, incase an error like this happens again.